refactor: FilterOptions 비즈니스 로직 분리
- filterOptions: useFilterOptions hook
